requirements.txt
mdtex2html
pypinyin
tiktoken
socksio
tqdm
colorama
Pygments
llama_index
langchain~=0.0.191
markdown
PyPDF2
pdfplumber
pandas
commentjson
openpyxl
python-dotenv~=1.0.0
bs4~=0.0.1
flask~=2.3.2
websockets~=11.0.3
requests~=2.30.0
beautifulsoup4~=4.12.2
Werkzeug~=2.3.4